requirements.txt
av==12.3.0
ffmpeg-python==0.2.0
numpy>=1.26.0
PyYAML>=6.0